+++ Source/SwiftLintFramework/Extensions/SwiftDeclarationKind+SwiftLint.swift
@@ -2,7 +2,7 @@
 //  SwiftDeclarationKind+SwiftLint.swift
 //  SwiftLint
 //
-//  Created by JP Simard on 2015-11-17.
+//  Created by JP Simard on 11/17/15.
 //  Copyright © 2015 Realm. All rights reserved.
 //