+++ Source/SwiftLintFramework/Rules/ConditionalReturnsOnNewline.swift
@@ -2,7 +2,7 @@
 //  ConditionalReturnsOnNewline.swift
 //  SwiftLint
 //
-//  Created by Rohan Dhaimade on 12/08/2016.
+//  Created by Rohan Dhaimade on 12/8/16.
 //  Copyright © 2016 Realm. All rights reserved.
 //