+++ Source/SwiftLintFramework/Rules/ControlStatementRule.swift
@@ -3,7 +3,7 @@
 //  SwiftLint
 //
 //  Created by Andrea Mazzini on 26/05/15.
-//  Copyright (c) 2015 Realm. All rights reserved.
+//  Copyright © 2015 Realm. All rights reserved.
 //
 
 import SourceKittenFramework
@@ -63,11 +63,11 @@ public struct ControlStatementRule: ConfigurationProviderRule {
                 : "\(statementKind)\\s*\\([^,{]*\\)\\s*\\{"
             return file.matchPattern(pattern).flatMap { match, syntaxKinds in
                 let matchString = file.contents.substring(match.location, length: match.length)
-                if self.isFalsePositive(matchString, syntaxKind: syntaxKinds.first) {
+                if isFalsePositive(matchString, syntaxKind: syntaxKinds.first) {
                     return nil
                 }
                 return StyleViolation(ruleDescription: type(of: self).description,
-                    severity: self.configuration.severity,
+                    severity: configuration.severity,
                     location: Location(file: file, characterOffset: match.location))
             }
         }