+++ Source/SwiftLintFramework/Rules/LegacyConstantRule.swift
@@ -2,7 +2,7 @@
 //  LegacyConstantRule.swift
 //  SwiftLint
 //
-//  Created by Aaron McTavish on 12/01/2016.
+//  Created by Aaron McTavish on 12/1/16.
 //  Copyright © 2016 Realm. All rights reserved.
 //