+++ Source/SwiftLintFramework/Rules/ValidIBInspectableRule.swift
@@ -2,7 +2,7 @@
 //  ValidIBInspectableRule.swift
 //  SwiftLint
 //
-//  Created by Marcelo Fabri on 10/20/2016.
+//  Created by Marcelo Fabri on 10/20/16.
 //  Copyright © 2016 Realm. All rights reserved.
 //